remove useless module
